+++ src/test/index.test.ts
@@ -22,30 +22,46 @@ test.afterEach.always(() => {
   sinon.restore()
 })
 
-test.serial('should log correctly with no parameter', async (t) => {
+test.serial('should log correctly with no message parameter', async (t) => {
   const logger = new LevelLogger()
   logger.info()
   calledOnceWithExactly(t, t.context.consoleInfoStub, '')
 })
 
-test.serial('should log correctly with a single parameter', async (t) => {
+test.serial('should log correctly with a single message parameter', async (t) => {
   const logger = new LevelLogger()
   logger.info('Hello world!')
   calledOnceWithExactly(t, t.context.consoleInfoStub, 'Hello world!')
 })
 
-test.serial('should log correctly with multiple parameters', async (t) => {
+test.serial('should log correctly with multiple message parameters', async (t) => {
   const logger = new LevelLogger()
   logger.info('Hello', 'again,', 'world!')
   calledOnceWithExactly(t, t.context.consoleInfoStub, 'Hello again, world!')
 })
 
+test.serial('should log nothing when no message parameter with some prefixes', async (t) => {
+  const logger = new LevelLogger({
+    prefixes: ['A', 'B']
+  })
+  logger.info()
+  calledOnceWithExactly(t, t.context.consoleInfoStub, '')
+})
+
+test.serial('should log correctly some prefixes and multiple message parameters', async (t) => {
+  const logger = new LevelLogger({
+    prefixes: ['A', 'B']
+  })
+  logger.info('MessageC', 'MessageD')
+  calledOnceWithExactly(t, t.context.consoleInfoStub, 'A B MessageC MessageD')
+})
+
 test.serial('should resolve timestamp prefix', async (t) => {
-  const timestampRegex = /\d{4}-\d{2}-\d{2} \d{2}:\d{2}:\d{2}/
+  const timestampRegex = /^\d{4}-\d{2}-\d{2} \d{2}:\d{2}:\d{2} message$/
   const logger = new LevelLogger({
     prefixes: [LogTags.TIMESTAMP]
   })
-  logger.info()
+  logger.info('message')
   const args = t.context.consoleInfoStub.args
   t.deepEqual(args.length, 1)
   t.deepEqual(args[0].length, 1)
@@ -53,11 +69,11 @@ test.serial('should resolve timestamp prefix', async (t) => {
 })
 
 test.serial('should resolve ISO timestamp prefix', async (t) => {
-  const timestampRegex = /\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}\.\d{3}Z/
+  const timestampRegex = /^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}\.\d{3}Z message$/
   const logger = new LevelLogger({
     prefixes: [LogTags.ISO_TIMESTAMP]
   })
-  logger.info()
+  logger.info('message')
   const args = t.context.consoleInfoStub.args
   t.deepEqual(args.length, 1)
   t.deepEqual(args[0].length, 1)
@@ -68,8 +84,8 @@ test.serial('should resolve log level prefix', async (t) => {
   const logger = new LevelLogger({
     prefixes: [LogTags.MESSAGE_LEVEL]
   })
-  logger.info()
-  calledOnceWithExactly(t, t.context.consoleInfoStub, 'INFO')
+  logger.info('message')
+  calledOnceWithExactly(t, t.context.consoleInfoStub, 'INFO message')
 })
 
 test.serial('extend should retain previous logger options', async (t) => {
@@ -83,15 +99,12 @@ test.serial('extend should retain previous logger options', async (t) => {
 })
 
 test.serial('extend should allow change of prefixes', async (t) => {
-  let logger = new LevelLogger({
+  let loggerA = new LevelLogger({
     prefixes: ['A']
   })
-  logger.info()
-  logger = logger.extend({
+  let loggerB = loggerA.extend({
     prefixes: ['B']
   })
-  logger.info()
-  t.true(t.context.consoleInfoStub.calledTwice)
-  t.deepEqual(t.context.consoleInfoStub.firstCall.args, ['A'])
-  t.deepEqual(t.context.consoleInfoStub.secondCall.args, ['B'])
+  t.deepEqual(loggerA.prefixes, ['A'])
+  t.deepEqual(loggerB.prefixes, ['B'])
 })